Implemented saving/checking account service specific methods. Added JUnit tests.

+++ src/main/java/hr/banking/domain/Account.java
@@ -16,9 +16,9 @@ public class Account {
 
     private Owner owner;
 
-    private Long balance;
+    private Float balance;
 
-    public Account(Owner owner, Long balance){
+    public Account(Owner owner, Float balance){
         this.owner = owner;
         this.balance = balance;
     }
